Move stuff into ouster pacukage
