+++ packages/nominal-ouster/nominal/ouster/__init__.py
@@ -1,3 +1,4 @@
-from nominal.thirdparty.ouster._convert import convert_ouster_dataset
+from nominal.ouster._convert import convert_ouster_dataset
+from nominal.ouster.spatial import upload_point_cloud
 
-__all__ = ["convert_ouster_dataset"]
+__all__ = ["convert_ouster_dataset", "upload_point_cloud"]
